The size of the `trap::Info` field has a significant impact on the size
of `tracer::Item`, which gets moved a lot. These changes reduce the size
of `trap::Info` by using a smaller integer type for `ecause`.

+++ src/types/trap.rs
@@ -6,7 +6,7 @@
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct Info {
     /// Cause of the trap or interrupt
-    pub ecause: u64,
+    pub ecause: u16,
     /// Appropriate `utval`/`stval`/`vstval`/`mtval`
     ///
     /// This field also indicates whether the trap is an interrupt (`None`) or
